+++ src/main/java/tarehart/rlbot/intercept/strike/DribbleStrike.kt
@@ -12,7 +12,7 @@ import tarehart.rlbot.time.Duration
 class DribbleStrike: StrikeProfile() {
 
     override val preDodgeTime = Duration.ZERO
-    override val speedBoost = 0.0
+    override val speedBoost = 0.0F
     override val postDodgeTime = Duration.ZERO
     override val style = Style.DRIBBLE
     override val isForward = false
@@ -25,11 +25,11 @@ class DribbleStrike: StrikeProfile() {
         return intercept.space.z <= MAX_HEIGHT_OF_BALL_FOR_DRIBBLE
     }
 
-    override fun getPreKickWaypoint(car: CarData, intercept: Intercept, desiredKickForce: Vector3, expectedArrivalSpeed: Double): PreKickWaypoint? {
+    override fun getPreKickWaypoint(car: CarData, intercept: Intercept, desiredKickForce: Vector3, expectedArrivalSpeed: Float): PreKickWaypoint? {
         return AnyFacingPreKickWaypoint(
                 position = intercept.space.flatten(),
                 idealFacing = desiredKickForce.flatten(),
-                allowableFacingError = Math.PI,
+                allowableFacingError = Math.PI.toFloat(),
                 expectedTime = intercept.time)
     }
