+++ src/main/java/tarehart/rlbot/math/Polygon.kt
@@ -19,7 +19,7 @@ class Polygon(points: Array<Vector2>) {
             throw IllegalArgumentException("Field area must have at least 3 points!")
         }
 
-        val path = Path2D.Double()
+        val path = Path2D.Float()
         path.moveTo(points[0].x, points[0].y)
         for (i in 1 until points.size) {
             path.lineTo(points[i].x, points[i].y)
@@ -29,6 +29,6 @@ class Polygon(points: Array<Vector2>) {
     }
 
     operator fun contains(test: Vector2): Boolean {
-        return area.contains(test.x, test.y)
+        return area.contains(test.x.toDouble(), test.y.toDouble())
     }
 }
