+++ src/main/java/tarehart/rlbot/routing/PositionFacing.kt
@@ -1,11 +1,10 @@
 package tarehart.rlbot.routing
 
 import tarehart.rlbot.input.CarData
-import tarehart.rlbot.math.VectorUtil
 import tarehart.rlbot.math.vector.Vector2
 import tarehart.rlbot.ui.ArenaDisplay
-
-import java.awt.*
+import java.awt.Color
+import java.awt.Graphics2D
 
 
 class PositionFacing(val position: Vector2, facing: Vector2) {
@@ -16,10 +15,10 @@ class PositionFacing(val position: Vector2, facing: Vector2) {
 
     fun drawDebugInfo(graphics: Graphics2D) {
         graphics.color = Color.green
-        ArenaDisplay.drawCar(this, 0.0, graphics)
+        ArenaDisplay.drawCar(this, 0F, graphics)
     }
 
-    fun distanceRemaining(entity: Vector2): Double {
+    fun distanceRemaining(entity: Vector2): Float {
         val toPosition = position - entity
         return toPosition.dotProduct(facing) / facing.magnitudeSquared()
     }
